64 bit compat, handle missing textures and extend bvh tester

+++ src/mesh.h
@@ -30,10 +30,10 @@ public:
 	u32 getBvhRootNode() const override { return _bvh_root_node; };
 
 	bool isDynamic() const override { return false; };
-	u32 maxNumVertices() const override { return _vertices.size(); };
-	u32 maxNumTriangles() const override { return _triangles.size(); };
-	u32 maxNumMaterials() const override { return _materials.size(); };
-	u32 maxNumBvhNodes() const override { return _bvh_nodes.size(); };
+	u32 maxNumVertices() const override { return (u32)_vertices.size(); };
+	u32 maxNumTriangles() const override { return (u32)_triangles.size(); };
+	u32 maxNumMaterials() const override { return (u32)_materials.size(); };
+	u32 maxNumBvhNodes() const override { return (u32)_bvh_nodes.size(); };
 	void buildBvh() override { };// Only necessary for dynamic objects
 private:
 	void addSubMesh(const aiScene* scene, uint mesh_index, const glm::mat4& transform_matrix, const char* texturePath);
